# docs/openapi/lmeval-eval-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LMEval Evaluation API",
    "version": "1.0.0",
    "description": "Machine-readable contract for the local LMEval API. Paths are relative to the deployment base path. Delete, git commit/revert, and promotion actions require separate explicit authorization."
  },
  "servers": [{ "url": "/api/eval" }],
...
  "paths": {
...
    "/evaluations": {
      "get": { "tags": ["Evaluations"], "summary": "List evaluations", "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string" } }, { "name": "promptId", "in": "query", "schema": { "type": "string" } }, { "name": "modelId", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "$ref": "#/components/responses/Json" } } },
      "post": { "tags": ["Evaluations"], "summary": "Backward-compatible create and start", "requestBody": { "$ref": "#/components/requestBodies/EvaluationInput" }, "responses": { "202": { "description": "Pending evaluation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationConfig" } } } }, "400": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } }
    },
    "/evaluations/validate": { "post": { "tags": ["Evaluations"], "summary": "Validate without persisting", "requestBody": { "$ref": "#/components/requestBodies/EvaluationInput" }, "responses": { "200": { "description": "Validation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Validation" } } } }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/drafts": { "post": { "tags": ["Evaluations"], "summary": "Create a reviewable draft without model calls", "requestBody": { "$ref": "#/components/requestBodies/EvaluationInput" }, "responses": { "201": { "$ref": "#/components/responses/Draft" }, "400": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
...
    "/evaluations/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Evaluations"], "summary": "Get evaluation configuration", "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Error" } } },
      "patch": { "tags": ["Evaluations"], "summary": "Patch a draft", "requestBody": { "$ref": "#/components/requestBodies/EvaluationPatch" }, "responses": { "200": { "$ref": "#/components/responses/Draft" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } },
      "delete": { "tags": ["Evaluations"], "summary": "Permanently delete an evaluation", "x-destructive": true, "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/evaluations/{id}/run": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "post": { "tags": ["Evaluations"], "summary": "Start a draft exactly once", "responses": { "202": { "$ref": "#/components/responses/Draft" }, "409": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/feedback": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "tags": ["Evaluations"], "summary": "Pollable durable orchestration and recovery snapshot", "responses": { "200": { "description": "Feedback snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Feedback" } } } }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/resume": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "post": { "tags": ["Evaluations"], "summary": "Resume unfinished exact cells in the same evaluation", "description": "Performs a read-only compatibility/catalog preflight, creates a durable attempt, reuses valid checkpoints, and returns before remaining provider work completes.", "responses": { "202": { "description": "Durable resume attempt accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResumeResponse" } } } }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/results": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "tags": ["Evaluations"], "summary": "Detailed result cells", "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/summary": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "tags": ["Evaluations"], "summary": "Aggregate summary", "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/testcases": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "tags": ["Evaluations"], "summary": "Resolved cases actually run", "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/evaluations/{id}/history": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "tags": ["Evaluations"], "summary": "Related evaluation history", "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/evaluations/{id}/regression": { "parameters": [{ "$ref": "#/components/parameters/Id" }, { "name": "baselineSlug", "in": "query", "required": true, "schema": { "type": "string" } }], "get": { "tags": ["Evaluations"], "summary": "Compare with a saved baseline", "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/cancel": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "post": { "tags": ["Evaluations"], "summary": "Request cancellation without deleting checkpoints", "responses": { "202": { "$ref": "#/components/responses/Json" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/evaluations/{id}/retry": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "post": { "tags": ["Evaluations"], "summary": "Create a retry evaluation", "responses": { "202": { "$ref": "#/components/responses/Json" } } } },
    "/evaluations/{id}/summary-analysis": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "get": { "tags": ["Evaluations"], "summary": "Get cached summary analysis", "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Error" } } }, "post": { "tags": ["Evaluations"], "summary": "Generate summary analysis", "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/evaluations/{id}/export": { "parameters": [{ "$ref": "#/components/parameters/Id" }, { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["html", "md"] } }], "get": { "tags": ["Evaluations"], "summary": "Export report", "responses": { "200": { "description": "HTML or Markdown report" } } } },
    "/evaluations/{id}/baseline": { "parameters": [{ "$ref": "#/components/parameters/Id" }], "post": { "tags": ["Evaluations"], "summary": "Save summary as baseline", "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
...
  },
  "components": {
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "requestBodies": {
      "EvaluationInput": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationInput" }, "example": { "name": "Classification smoke", "promptIds": ["prompt-1"], "modelIds": ["local::model-a"], "comparisonMode": "matrix", "inlineTestCases": [{ "id": "case-1", "userMessage": "<memory>Example</memory>", "expectedLabels": ["Note"] }], "runsPerCell": 1, "inference": { "temperature": 0.3, "maxTokens": 1000 } } } } },
      "EvaluationPatch": { "required": true, "content": { "application/json": { "schema": { "type": "object", "description": "Any caller-controlled EvaluationInput fields; validation applies to the merged draft." } } } }
    },
    "responses": {
      "Json": { "description": "JSON response", "content": { "application/json": { "schema": {} } } },
      "Error": { "description": "Machine-readable error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "Evaluation configuration is invalid", "code": "EVALUATION_VALIDATION_FAILED" } } } },
      "Draft": { "description": "Evaluation resource and validation", "content": { "application/json": { "schema": { "type": "object", "required": ["evaluation", "validation", "browserPaths"], "properties": { "evaluation": { "$ref": "#/components/schemas/EvaluationConfig" }, "validation": { "$ref": "#/components/schemas/Validation" }, "browserPaths": { "$ref": "#/components/schemas/BrowserPaths" } } } } } }
    },
    "schemas": {
...
      "Inference": { "type": "object", "required": ["temperature", "maxTokens"], "properties": { "temperature": { "type": "number", "minimum": 0, "maximum": 2 }, "maxTokens": { "type": "integer", "minimum": 1, "maximum": 1000000 }, "seed": { "type": "integer", "minimum": -2147483648, "maximum": 2147483647 } } },
      "PromptVersion": { "type": "object", "required": ["promptId", "version"], "properties": { "promptId": { "type": "string" }, "version": { "type": "integer", "minimum": 1 } } },
      "EvaluationInput": { "type": "object", "required": ["name", "promptIds", "modelIds"], "properties": { "name": { "type": "string", "minLength": 1 }, "promptIds": { "type": "array", "minItems": 1, "items": { "type": "string" } }, "promptVersions": { "type": "array", "items": { "$ref": "#/components/schemas/PromptVersion" } }, "modelIds": { "type": "array", "minItems": 1, "items": { "type": "string", "pattern": "^.+::.+$" } }, "comparisonMode": { "type": "string", "enum": ["model", "prompt", "matrix"] }, "purposeTemplateId": { "type": "string" }, "testSuiteId": { "type": "string" }, "userMessage": { "type": "string" }, "inlineTestCases": { "type": "array", "items": { "type": "object" } }, "templateId": { "type": "string" }, "judgeModelId": { "type": "string" }, "enablePairwise": { "type": "boolean" }, "runsPerCell": { "type": "integer", "minimum": 1, "maximum": 10 }, "sessionId": { "type": "string" }, "sessionVersion": { "type": "integer" }, "inference": { "$ref": "#/components/schemas/Inference" }, "benchmarkMode": { "type": "string", "enum": ["calibration", "promotion-check"] }, "campaignId": { "type": "string" }, "campaignRole": { "type": "string", "enum": ["supplemental"] } } },
      "EvaluationConfig": { "allOf": [{ "$ref": "#/components/schemas/EvaluationInput" }, { "type": "object", "required": ["id", "status", "createdAt", "updatedAt"], "properties": { "id": { "type": "string" }, "status": { "type": "string", "enum": ["draft", "pending", "running", "interrupted", "completed", "failed", "cancelled"] }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } }] },
      "ValidationIssue": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string" }, "field": { "type": "string" }, "message": { "type": "string" } } },
      "Validation": { "type": "object", "required": ["valid", "errors", "warnings"], "properties": { "valid": { "type": "boolean" }, "errors": { "type": "array", "items": { "$ref": "#/components/schemas/ValidationIssue" } }, "warnings": { "type": "array", "items": { "$ref": "#/components/schemas/ValidationIssue" } } } },
      "BrowserPaths": { "type": "object", "required": ["config", "run", "results", "summary"], "properties": { "config": { "type": "string" }, "run": { "type": "string" }, "results": { "type": "string" }, "summary": { "type": "string" } } },
      "Recovery": { "type": "object", "required": ["state", "reusedCells", "remainingCells", "requiresModelCalls"], "properties": { "state": { "type": "string", "enum": ["not-needed", "eligible", "finalization-only", "blocked", "legacy-unrecoverable"] }, "reasonCode": { "type": "string" }, "reusedCells": { "type": "integer" }, "remainingCells": { "type": "integer" }, "lastAttemptId": { "type": "string" }, "requiresModelCalls": { "type": "boolean" } } },
      "Feedback": { "type": "object", "required": ["evalId", "status", "progress", "validation", "failures", "readiness", "appBasePath", "browserPaths"], "properties": { "evalId": { "type": "string" }, "status": { "type": "string" }, "progress": { "type": "object", "description": "Includes checkpoint-derived durablyCommitted and remaining for schema-v1 runs." }, "validation": { "$ref": "#/components/schemas/Validation" }, "failures": { "type": "array", "items": { "type": "object" } }, "readiness": { "type": "object" }, "verdict": { "type": ["object", "null"] }, "attempt": { "type": "object" }, "recovery": { "$ref": "#/components/schemas/Recovery" }, "appBasePath": { "type": "string" }, "browserPaths": { "$ref": "#/components/schemas/BrowserPaths" } } },
      "ResumeResponse": { "type": "object", "required": ["evaluationId", "attemptId", "mode", "counts", "preflight", "browserPaths"], "properties": { "evaluationId": { "type": "string" }, "attemptId": { "type": "string" }, "resumedFromAttemptId": { "type": "string" }, "mode": { "type": "string", "enum": ["execute-remaining", "finalization-only"] }, "counts": { "type": "object" }, "preflight": { "type": "object" }, "browserPaths": { "$ref": "#/components/schemas/BrowserPaths" } } },
      "Error": { "type": "object", "required": ["error", "code"], "properties": { "error": { "type": "string" }, "code": { "type": "string" }, "reasonCode": { "type": "string" }, "checks": { "type": "array", "items": { "type": "object" } }, "validation": { "$ref": "#/components/schemas/Validation" } } }
    }
  }
}